Revised and complete thesis project
This is the code used for the thesis to generate different RAG pipelines for a chatbot, trying to optimize and analyze different steps while taking part of the SATO research on LLM as a jury. Without the vectorstores and the private company data the code cannot be run but can be used as reference for other similar implementations.